+++ QuadFinal/PIDcontrol.h
@@ -19,20 +19,20 @@ class PIDController
 {
 public:
 	PIDController();
-	void changeGain(int p, int i, int d);
+	void changeGain(float p, float i, float d);
 	~PIDController();
 	void move(int deltaPos);
-	void setSetPoint(int newPoint);
+	void setSetPoint(float newPoint);
 
-	int PID(int currentValue);
-	int PID(int currentValue, int dTerm);
+	float PID(float currentValue);
+	float PID(float currentValue, float dTerm);
 private:
-	int setPoint;
-	int integralSumTerm;
-	int lastValue;
-	int kP, kD, kI;
-	int deltaTime, previousMillis;
-	int dTerm, error;
+	float setPoint;
+	float integralSumTerm;
+	float lastValue;
+	float kP, kD, kI;
+	float deltaTime, previousMillis;
+	float dTerm, error;
 };
 
 #endif
